Support promises and returning matcher from test
Make the chained methods always return a promise object, that also has
all the matchers defined on it.

Jest accepts this as the return value from a test because it is a Promise.

Also, if any intermediate step returns a Promise, the promise returned by
jest-chain will wait for that Promise.